convert int to uint32_t

+++ source/games/Battleships.cpp
@@ -22,9 +22,9 @@ void Battleships::Setup_Game()
     m_MovesRemainingOne.clear();
     m_MovesRemainingTwo.clear();
 
-    for (int i = 0, GridNumber = 0; i < 10; i++)
+    for (uint32_t i = 0, GridNumber = 0; i < 10; i++)
     {
-        for (int j = 0; j < 10; j++, GridNumber++)
+        for (uint32_t j = 0; j < 10; j++, GridNumber++)
         {
             m_BoardOne[i][j] = ' ';
             m_BoardTwo[i][j] = ' ';
@@ -140,18 +140,18 @@ void Battleships::Get_User_Ship_Positions()
             "Please enter the 2 grid locations for the Patrol Boat"
     };
     std::array<char, 5> ShipLetters = {'C', 'B', 'D', 'S', 'P'};
-    std::array<int, 5> ShipSizes = {5, 4, 3, 3, 2};
-    int lastShipRow = 0, lastShipColumn = 0;
+    std::array<uint32_t, 5> ShipSizes = {5, 4, 3, 3, 2};
+    uint32_t lastShipRow = 0, lastShipColumn = 0;
 
     m_Terminal.Set_Cursor_Visibility(true);
 
-    for (int i = 0; i < 5; i++) // for each ship
+    for (uint32_t i = 0; i < 5; i++) // for each ship
     {
-        std::vector<int> ShipRows, ShipColumns;
+        std::vector<uint32_t> ShipRows, ShipColumns;
         bool ShipOrientationHorizontal = false;
         bool ShipOrientationVertical = false;
 
-        for (int j = 0; j < ShipSizes[i]; j++) // for each ship grid locations
+        for (uint32_t j = 0; j < ShipSizes[i]; j++) // for each ship grid locations
         {
             std::string Output = Get_Game_Display() + m_StringBuilder.New_Line_Centred(ShipInstructions[i]);
             Output += m_StringBuilder.Empty_Line() + m_StringBuilder.Empty_Line() + m_StringBuilder.Empty_Line() + m_StringBuilder.Empty_Line();
@@ -159,7 +159,7 @@ void Battleships::Get_User_Ship_Positions()
 
             m_Terminal.Output_To_Terminal(Output);
 
-            int KeyPress, Row, Column;
+            uint32_t KeyPress, Row, Column;
 
             if (j == 0)
             {
@@ -252,7 +252,7 @@ void Battleships::Get_User_Ship_Positions()
                         if ((RowsSame && ColumnsSame) || (!RowsSame && !ColumnsSame))
                             continue;
 
-                        if (RowsSame && (std::abs(Column - ShipColumns.back()) == 1))
+                        if (RowsSame && (std::abs((int)Column - (int)ShipColumns.back()) == 1))
                         {
                             if (ShipRows.size() == 1)
                             {
@@ -269,7 +269,7 @@ void Battleships::Get_User_Ship_Positions()
                             }
                         }
 
-                        else if (ColumnsSame && (std::abs(Row - ShipRows.back()) == 1))
+                        else if (ColumnsSame && (std::abs((int)Row - (int)ShipRows.back()) == 1))
                         {
                             if (ShipRows.size() == 1)
                             {
@@ -299,7 +299,7 @@ void Battleships::Get_User_Ship_Positions()
         }
 
         // All grid location received and are valid, so place ship
-        for (unsigned int j = 0; j < ShipRows.size(); j++)
+        for (uint32_t j = 0; j < ShipRows.size(); j++)
             m_BoardOne[ShipRows[j]][ShipColumns[j]] = ShipLetters[i];
     }
 
@@ -308,15 +308,15 @@ void Battleships::Get_User_Ship_Positions()
 
 void Battleships::Get_AI_Ship_Positions(std::array<std::array<char, 10>, 10> &AIBoard)
 {
-    std::array<int, 5> ShipSizes = {5, 4, 3, 3, 2};
+    std::array<uint32_t, 5> ShipSizes = {5, 4, 3, 3, 2};
     std::array<char, 5> ShipLetters = {'C', 'B', 'D', 'S', 'P'};
 
-    for (int i = 0; i < 5; i++)
+    for (uint32_t i = 0; i < 5; i++)
     {
         while (true)
         {
-            std::vector<unsigned int> ShipRows, ShipColumns;
-            unsigned int Row, Column;
+            std::vector<uint32_t> ShipRows, ShipColumns;
+            uint32_t Row, Column;
 
             // Get a random ship position
             if ((m_RandomNumberGenerator() % 2) == 0) // Horizontal
@@ -328,7 +328,7 @@ void Battleships::Get_AI_Ship_Positions(std::array<std::array<char, 10>, 10> &AI
                 Row = m_RandomNumberGenerator() % 10; // Any row number allowed as ship is horizontal
 
                 // Horizontal positions have a difference in columns of 1
-                for (int j = 0; j < ShipSizes[i]; j++)
+                for (uint32_t j = 0; j < ShipSizes[i]; j++)
                 {
                     ShipRows.push_back(Row);
                     ShipColumns.push_back(Column + j);
@@ -344,7 +344,7 @@ void Battleships::Get_AI_Ship_Positions(std::array<std::array<char, 10>, 10> &AI
                 Row = m_RandomNumberGenerator() % (11 - ShipSizes[i]);
 
                 // Vertical positions have a difference in rows of 1
-                for (int j = 0; j < ShipSizes[i]; j++)
+                for (uint32_t j = 0; j < ShipSizes[i]; j++)
                 {
                     ShipRows.push_back(Row + j);
                     ShipColumns.push_back(Column);
@@ -353,7 +353,7 @@ void Battleships::Get_AI_Ship_Positions(std::array<std::array<char, 10>, 10> &AI
 
             // Check if ship is already located on requested ship position
             bool Errors = false;
-            for (unsigned int j = 0; j < ShipRows.size(); j++)
+            for (uint32_t j = 0; j < ShipRows.size(); j++)
                 if (AIBoard[ShipRows[j]][ShipColumns[j]] == 'C' ||
                     AIBoard[ShipRows[j]][ShipColumns[j]] == 'B' ||
                     AIBoard[ShipRows[j]][ShipColumns[j]] == 'D' ||
@@ -363,7 +363,7 @@ void Battleships::Get_AI_Ship_Positions(std::array<std::array<char, 10>, 10> &AI
 
             if (!Errors) // Place ship
             {
-                for (unsigned int j = 0; j < ShipRows.size(); j++)
+                for (uint32_t j = 0; j < ShipRows.size(); j++)
                     AIBoard[ShipRows[j]][ShipColumns[j]] = ShipLetters[i];
                 break;
             }
@@ -374,14 +374,14 @@ void Battleships::Get_AI_Ship_Positions(std::array<std::array<char, 10>, 10> &AI
 bool Battleships::Game_Over()
 {
     bool PlayerOneShipsPresent = false;
-    for (int i = 0; i < 10; i++)
-        for (int j = 0; j < 10; j++)
+    for (uint32_t i = 0; i < 10; i++)
+        for (uint32_t j = 0; j < 10; j++)
             if (m_BoardOne[i][j] == 'C' || m_BoardOne[i][j] == 'B' || m_BoardOne[i][j] == 'D' || m_BoardOne[i][j] == 'S' || m_BoardOne[i][j] == 'P')
                 PlayerOneShipsPresent = true;
 
     bool PlayerTwoShipsPresent = false;
-    for (int i = 0; i < 10; i++)
-        for (int j = 0; j < 10; j++)
+    for (uint32_t i = 0; i < 10; i++)
+        for (uint32_t j = 0; j < 10; j++)
             if (m_BoardTwo[i][j] == 'C' || m_BoardTwo[i][j] == 'B' || m_BoardTwo[i][j] == 'D' || m_BoardTwo[i][j] == 'S' || m_BoardTwo[i][j] == 'P')
                 PlayerTwoShipsPresent = true;
 
@@ -409,7 +409,7 @@ void Battleships::Execute_Next_User_Command()
 
     m_Terminal.Output_To_Terminal(Output);
 
-    int KeyPress, Command, Row = m_PreviousCommand / 10, Column = m_PreviousCommand % 10;
+    uint32_t KeyPress, Command, Row = m_PreviousCommand / 10, Column = m_PreviousCommand % 10;
 
     m_Terminal.Set_Cursor_Visibility(true);
 
@@ -477,11 +477,11 @@ void Battleships::Execute_Next_AI_Command()
 }
 
 void Battleships::Execute_Command(std::array<std::array<char, 10>, 10> &OpponentBoard,
-                                  std::unordered_map<char, int> &OpponentShipsRemaining,
-                                  std::vector<int> &MovesRemaining,
-                                  const int &Command)
+                                  std::unordered_map<char, uint32_t> &OpponentShipsRemaining,
+                                  std::vector<uint32_t> &MovesRemaining,
+                                  const uint32_t &Command)
 {
-    int Row = Command / 10, Column = Command % 10;
+    uint32_t Row = Command / 10, Column = Command % 10;
 
     if (OpponentBoard[Row][Column] == 'C' || OpponentBoard[Row][Column] == 'B' || OpponentBoard[Row][Column] == 'D' || OpponentBoard[Row][Column] == 'S' || OpponentBoard[Row][Column] == 'P')
     {
@@ -521,11 +521,11 @@ std::string Battleships::Get_Game_Display()
     Output += m_StringBuilder.New_Line_Left_Justified("                    Player One                                                                                       Player Two");
 
     Output += (char)186 + std::string("   ");
-    for (int i = 0; i < 2; i++)
+    for (uint32_t i = 0; i < 2; i++)
     {
         Output += (char)218;
         Output.insert(Output.size(), 3, (char)196);
-        for (int j = 0; j < 10; j++)
+        for (uint32_t j = 0; j < 10; j++)
         {
             Output += (char)194;
             Output.insert(Output.size(), 3, (char)196);
@@ -538,7 +538,7 @@ std::string Battleships::Get_Game_Display()
     Output += std::string("   ") + (char)186 + "\n";
 
     Output += (char)186 + std::string("   ");
-    for (int i = 0; i < 2; i++)
+    for (uint32_t i = 0; i < 2; i++)
     {
         Output += (char)179 + std::string("   ") + (char)179 + " A " + (char)179 + " B " + (char)179 + " C " + (char)179 + " D " + (char)179 + " E " + (char)179 + " F " + (char)179 + " G " + (char)179 + " H " + (char)179 + " I " + (char)179 + " J " + (char)179;
 
@@ -548,7 +548,7 @@ std::string Battleships::Get_Game_Display()
     Output += std::string("   ") + (char)186 + "\n";
 
     // Main parts of both boards and centre information
-    for (int i = 0; i < 10; i++)
+    for (uint32_t i = 0; i < 10; i++)
     {
         // First Line
         // Left outer box edge
@@ -557,7 +557,7 @@ std::string Battleships::Get_Game_Display()
         // Player One Board horizontal dividers
         Output += (char)195;
         Output.insert(Output.size(), 3, (char)196);
-        for (int j = 0; j < 10; j++)
+        for (uint32_t j = 0; j < 10; j++)
         {
             Output += (char)197;
             Output.insert(Output.size(), 3, (char)196);
@@ -570,7 +570,7 @@ std::string Battleships::Get_Game_Display()
         else if (i == 5) // Battleship
         {
             Output += "   ";
-            for (int j = 0; j < 4; j++, Output += " ")
+            for (uint32_t j = 0; j < 4; j++, Output += " ")
             {
                 if (j < m_ShipsRemainingOne.at('B'))
                     Output.insert(Output.size(), 3, (char)178);
@@ -578,7 +578,7 @@ std::string Battleships::Get_Game_Display()
                     Output.insert(Output.size(), 3, (char)176);
             }
             Output += "              ";
-            for (int j = 0; j < 4; j++, Output += " ")
+            for (uint32_t j = 0; j < 4; j++, Output += " ")
             {
                 if (j < (4 - m_ShipsRemainingTwo.at('B')))
                     Output.insert(Output.size(), 3, (char)176);
@@ -592,7 +592,7 @@ std::string Battleships::Get_Game_Display()
         else if (i == 8) // Submarine
         {
             Output += "   ";
-            for (int j = 0; j < 3; j++, Output += " ")
+            for (uint32_t j = 0; j < 3; j++, Output += " ")
             {
                 if (j < m_ShipsRemainingOne.at('S'))
                     Output.insert(Output.size(), 3, (char)178);
@@ -600,7 +600,7 @@ std::string Battleships::Get_Game_Display()
                     Output.insert(Output.size(), 3, (char)176);
             }
             Output += "                      ";
-            for (int j = 0; j < 3; j++, Output += " ")
+            for (uint32_t j = 0; j < 3; j++, Output += " ")
             {
                 if (j < (3 - m_ShipsRemainingTwo.at('S')))
                     Output.insert(Output.size(), 3, (char)176);
@@ -617,7 +617,7 @@ std::string Battleships::Get_Game_Display()
         // Player Two Board horizontal dividers
         Output += (char)195;
         Output.insert(Output.size(), 3, (char)196);
-        for (int j = 0; j < 10; j++)
+        for (uint32_t j = 0; j < 10; j++)
         {
             Output += (char)197;
             Output.insert(Output.size(), 3, (char)196);
@@ -629,7 +629,7 @@ std::string Battleships::Get_Game_Display()
         Output += (char)186 + std::string("   ") + (char)179 + " " + std::to_string(i) + " " + (char)179;
 
         // Player One board ships
-        for (int j = 0; j < 10; j++)
+        for (uint32_t j = 0; j < 10; j++)
         {
             if (m_BoardOne[i][j] == 'C' || m_BoardOne[i][j] == 'B' || m_BoardOne[i][j] == 'D' || m_BoardOne[i][j] == 'S' || m_BoardOne[i][j] == 'P')
                 Output.insert(Output.size(), 3, (char)178);
@@ -650,7 +650,7 @@ std::string Battleships::Get_Game_Display()
         else if (i == 3) // Carrier
         {
             Output += "   ";
-            for (int j = 0; j < 5; j++, Output += " ")
+            for (uint32_t j = 0; j < 5; j++, Output += " ")
             {
                 if (j < m_ShipsRemainingOne.at('C'))
                     Output.insert(Output.size(), 3, (char)178);
@@ -658,7 +658,7 @@ std::string Battleships::Get_Game_Display()
                     Output.insert(Output.size(), 3, (char)176);
             }
             Output += "      ";
-            for (int j = 0; j < 5; j++, Output += " ")
+            for (uint32_t j = 0; j < 5; j++, Output += " ")
             {
                 if (j < (5 - m_ShipsRemainingTwo.at('C')))
                     Output.insert(Output.size(), 3, (char)176);
@@ -672,7 +672,7 @@ std::string Battleships::Get_Game_Display()
         else if (i == 6) // Destroyer
         {
             Output += "   ";
-            for (int j = 0; j < 3; j++, Output += " ")
+            for (uint32_t j = 0; j < 3; j++, Output += " ")
             {
                 if (j < m_ShipsRemainingOne.at('D'))
                     Output.insert(Output.size(), 3, (char)178);
@@ -680,7 +680,7 @@ std::string Battleships::Get_Game_Display()
                     Output.insert(Output.size(), 3, (char)176);
             }
             Output += "                      ";
-            for (int j = 0; j < 3; j++, Output += " ")
+            for (uint32_t j = 0; j < 3; j++, Output += " ")
             {
                 if (j < (3 - m_ShipsRemainingTwo.at('D')))
                     Output.insert(Output.size(), 3, (char)176);
@@ -694,7 +694,7 @@ std::string Battleships::Get_Game_Display()
         else if (i == 9) // Patrol Boat
         {
             Output += "   ";
-            for (int j = 0; j < 2; j++, Output += " ")
+            for (uint32_t j = 0; j < 2; j++, Output += " ")
             {
                 if (j < m_ShipsRemainingOne.at('P'))
                     Output.insert(Output.size(), 3, (char)178);
@@ -702,7 +702,7 @@ std::string Battleships::Get_Game_Display()
                     Output.insert(Output.size(), 3, (char)176);
             }
             Output += "                              ";
-            for (int j = 0; j < 2; j++, Output += " ")
+            for (uint32_t j = 0; j < 2; j++, Output += " ")
             {
                 if (j < (2 - m_ShipsRemainingTwo.at('P')))
                     Output.insert(Output.size(), 3, (char)176);
@@ -718,7 +718,7 @@ std::string Battleships::Get_Game_Display()
         Output += (char)179 + std::string(" ") + std::to_string(i) + " " + (char)179;
 
         // Player Two board ships
-        for (int j = 0; j < 10; j++)
+        for (uint32_t j = 0; j < 10; j++)
         {
             if ((m_GameOver || m_NumberOfPlayers == "0  ") && (m_BoardTwo[i][j] == 'C' || m_BoardTwo[i][j] == 'B' || m_BoardTwo[i][j] == 'D' || m_BoardTwo[i][j] == 'S' || m_BoardTwo[i][j] == 'P'))
                 Output.insert(Output.size(), 3, (char)178);
@@ -735,11 +735,11 @@ std::string Battleships::Get_Game_Display()
 
     // Bottom row of both boards
     Output += (char)186 + std::string("   ");
-    for (int i = 0; i < 2; i++)
+    for (uint32_t i = 0; i < 2; i++)
     {
         Output += (char)192;
         Output.insert(Output.size(), 3, (char)196);
-        for (int j = 0; j < 10; j++)
+        for (uint32_t j = 0; j < 10; j++)
         {
             Output += (char)193;
             Output.insert(Output.size(), 3, (char)196);
